Clean up sbom-graph-model and enrichment test/model code
Fix Defect.cvss type annotation from Optional[int] to Optional[float] to match actual CVSS score values
Use ProjectType enum values instead of raw strings for project.type in test fixtures
Remove unused imports across test files: pytest (enrichment tests), copy/json/MagicMock (CycloneDX tests), Any/call/RiskStatus (persistence tests), License/Project/Version (SPDX tests), DependencyVersion/HasVersion/License (conftest)

+++ sbom-graph-model/src/sbom_graph_model/model.py
@@ -88,7 +88,7 @@ class Defect:
         self.description: Optional[str] = None
         self.cwes: list[int] = []
         self.severity: Optional[str] = None
-        self.cvss: Optional[int] = None
+        self.cvss: Optional[float] = None
         self.cvss_string: Optional[str] = None
         self.source: Optional[tuple[str, str]] = None
         self.last_enriched_at: Optional[str] = None  # ISO timestamp
